update
add delete confirmation
add update option

+++ residentinfo.php
@@ -359,6 +359,7 @@ if (isset($_POST['fname'],$_POST['midname'],$_POST['lname'],$_POST['alias'],$_PO
                     
                     // Attempt select query execution
                     $sql = "SELECT * FROM residents";
+                    $confim = "Are you sure?";
                     if($result = $db->query($sql)){
                         if($result->rowCount() > 0){
                             echo '<table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">';
@@ -400,7 +401,7 @@ if (isset($_POST['fname'],$_POST['midname'],$_POST['lname'],$_POST['alias'],$_PO
                                         echo "<td>";
                                             echo '<a href="#id='. $row['ID'] .'" class="mr-3" title="View Record" data-toggle="tooltip"><span class="fa fa-eye"></span></a>';
                                             echo '<a href="#id='. $row['ID'] .'" class="mr-3" title="Update Record" data-toggle="tooltip"><span class="fa fa-pencil"></span></a>';
-                                            echo '<a href="delete.php?id='. $row['ID'] .'" title="Delete Record" data-toggle="tooltip" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
+                                            echo '<a href="delete.php?id='. $row['ID'] .'" onclick="return confirm(\'Are you sure to delete this Resident?\');" title="Delete Record" data-toggle="tooltip" data-toggle="tooltip"><span class="fa fa-trash"></span></a>';
                                         echo "</td>";
                                     echo "</tr>";
                                 }
@@ -1006,7 +1007,7 @@ if (isset($_POST['fname'],$_POST['midname'],$_POST['lname'],$_POST['alias'],$_PO
                 <select name ="year"class="form-select" id="select-year" required>
                   </select>     
             </div>
-            <div class="form-group">
+          <div class="form-group">
 						<label>Birthplace</label>
 						<input type="text" name = "birthplace"class="form-control" required>
 					</div>
